pyproject.toml
[project]
name = "ycncc-data-management"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pydantic[email]>=2.11.9",
    "pyyaml>=6.0.2",
    "ruff>=0.13.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["schemas"]

